ACE.Internal

Description

Internal Utility Functions for AC²E

Documentation

map' :: Text -> Double -> (String, Float)

Convert coerced map from

version :: String

Current AC²E version. Linked to https://github.com/electronics-and-drives/ace

mvnHome :: IO String

Get the default MAVEN home path

classPath :: IO String

Get the default CLASSPATH to AC²E from local MAVEN install

newtype ACEnv

Java AC²E Base Class

Constructors

ACEnv (J ('Class "edlab.eda.ace.AnalogCircuitEnvironment"))

newtype SEOEnv

Java AC²E Main Class

Constructors

SEOEnv (J ('Class "edlab.eda.ace.SingleEndedOpampEnvironment"))

newtype Pool

Java AC²E Environment Pool

Constructors

Pool (J ('Class "edlab.eda.ace.EnvironmentPool"))

mkEnv :: String -> String -> String -> IO SEOEnv

Contruct ACE Object

mkPool :: IO Pool

Create Empty ACE Pool

analyses :: SEOEnv -> IO [String]

Available Simulation Analyses

setSizing :: SEOEnv -> Map String Float -> IO ()

Set sizing paramters in Environment

currentPerformance :: SEOEnv -> IO (Map String Float)

Get current circuit performance

simulate' :: SEOEnv -> [String] -> [String] -> Map String Float -> IO (Map String Float)

Simulate Environment with given analyses blacklist, corners and sizing parameters

evaluate :: SEOEnv -> Map String Float -> IO (Map String Float)

Simulate Environment for default corners with empty black list

poolAdd :: Pool -> [SEOEnv] -> IO Pool

Add a list of Environments to Pool

evaluatePool :: [SEOEnv] -> [Map String Float] -> IO [Map String Float]

Simulate list of Environments in parallel

performanceIds :: SEOEnv -> IO [String]

Get available performance parameters

parameters :: SEOEnv -> IO [String]

All netlist Parameters

sizingParameters :: SEOEnv -> IO [String]

Sizing Parameters

initialSizing :: SEOEnv -> IO (Map String Float)

Initial Sizing Parameters for given Environment

randomSizing :: SEOEnv -> IO (Map String Float)

Random sizing paramters for given Environment